phase8-streamlit-deployment/requirements.txt
streamlit>=1.32.0
google-generativeai>=0.4.1
pandas>=2.2.0
python-dotenv>=1.0.1
